+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   commands to skip confirmation prompts (@vjik)
 - Chg #300: Replace deprecated `self::getDefaultName()` with `$this->getName()` (@Tigrov)
 - Enh #301: Add `MigrationBuilder::columnBuilder()` method (@Tigrov)
+- Enh #299: Update `MigrationBuilder::update()` method to adapt changes in `yiisoft/db` (@rustamwin)
 
 ## 1.2.0 November 27, 2024